requirements.txt
streamlit==1.45.0
transformers==4.51.3
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==2.2.5
python-dotenv==1.0.1